pyproject.toml
[project]
name = "resume-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-sqlite3>=0.0.1",
    "fastapi>=0.119.0",
    "pydantic-ai>=1.0.18",
    "yfinance>=0.2.66",
    "beautifulsoup4>=4.14.2",
    "logfire>=4.13.2",
    "fastapi>=0.119.0",
    "investpy>=1.0.8",
    "streamlit>=1.50.0",
    "pydantic-ai-slim[duckduckgo,huggingface]>=1.0.18",
    "torch>=2.9.0",
    "chromadb>=1.2.1",
]
